Crate data_blocks

data_blocks

Data blocks can be chained to provide verifiable assuredness that they contain network valid data and not injected.

This crate assumes these objects below are relevent to close_groups this allows integrity check of the DataChain in so far as it will hold valid data identifiers agreed by this group & all groups since the group/network started.

These chains allow nodes to become archive nodes on a network and also ensure data integrity AS LONG AS the data identifiers hold a validating element such as hash of the data itself.

Another purpose of these chains is to allow network restarts with valid data. Obviously this means the network nodes will have to tr and restart as the last known ID they had. Vaults will require to accept or reject such nodes in normal operation. On network restart though these nodes may be allowed to join a group if they can present a DataChain that appears healthy, even if there is not yet enough consensus to trust the data iself just yet. additional nodes will also join this group and hopefully confirm the data integrity is agreed as the last DataBlock should contain a majorit of existing group members that have signed.

Nodes do no require to become Archive nodes if they have limited bandwidth or disk space, but they are still valuable as transient nodes which may deliver data stored whle they are in the group. Such nodes may only be involved in consensus and routing stability messages, returning a Nack to any Get request due to upstream bandwidth limitations.

Several enhancement can be made to this scheme with a much deper investigation into any attack vectors. Such as

Structs

DataBlock
Used to validate chain linksi. On a network churn event the latest DataBlock is copied from the chain and sent To new node. The lost Nodes signature is removed. The new node receives this - signs a NodeBlock for thisDataIdentifierand returns it to thearchive node`

DataChain
Created by holder of chain, can be passed to others as proof of data held. This object is verifiable if : The last validation contains the majority of current close group OR on network restart the nodes all must try and restart on previous names. They can continue any validation of the holder of a chain. This requires nodes to always restart as last ID and if there was no restart they are rejected at vault level. If there was a restart then the nodes should validate and continue. N:B this means all nodes can use a named directory for data_store and clear if they restart as a new id. This allows cleanup of old data_cache directories.

NodeDataBlock
Sent by any group member when data is Put, Post or Delete in this group

Enums

DataIdentifier
Dummy data identifiers for this crate

Error
Error types.